I have completed the refactoring and documentation for the Argo generative models library. Here is a summary of the changes:
- **Code Refactoring**: The code from `argo/gen_models/api.py` has been moved to `argo/gen_models/__init__.py` to follow standard Python package practices. All imports have been updated accordingly.
- **Improved Documentation**:
  - The `GenerationTask` data class now has detailed documentation explaining the purpose of each parameter.
  - The `argo/gen_models/README.md` has been rewritten to reflect the new API and provide better examples.
  - A new main `README.md` file has been created with a project overview, a comprehensive example workflow, and suggestions for future development.
- **Standardized Input Validation**: Input validation has been standardized across all generative models to use `task.config.get()`, making the models more robust. A bug in the `F_RAGGenerator` where `task.scaffold` was not being used has also been fixed.
- **Test Refactoring**: The tests have been refactored into a new `tests` directory and converted to use the `pytest` framework. The tests are now organized by model, making them easier to maintain and run.
